feat: cursor finally movable feat: history added fix: console no longer crashes, when input is NULL
