+++ src/main.rs
@@ -1,73 +1,44 @@
-use std::io::{stdin, Write};
-use std::io::stdout;
-use std::process::{Child, Command, Stdio};
-use crate::built_ins::cd::execute_cd;
+use std::fmt::Error;
+use rustyline::{DefaultEditor};
+use rustyline::error::ReadlineError;
+use crate::command_handler::command_handler;
 use crate::config_handler::prompt::{read_prompt_statement_from_rsh, replace_placeholders};
 mod built_ins;
 mod config_handler;
+mod command_handler;
 
-fn main(){
+fn main() -> Result<(), Error> {
+    let mut rl = DefaultEditor::new().expect("Couldn't load Editor");
+
+    if rl.load_history(&dirs::home_dir().unwrap().join(".rsh_history")).is_err() {
+        println!("No history file found, will create one on exit")
+    }
     loop {
         if let Ok(print_statement) = read_prompt_statement_from_rsh() {
-            let replaced_prompt = replace_placeholders(&print_statement);
-            print!("{}", replaced_prompt);
-        }
-        stdout().flush().unwrap();
-
-        let mut input = String::new();
-        stdin().read_line(&mut input).unwrap();
-
-        // must be peekable so we know when we are on the last command
-        let mut commands = input.trim().split(" | ").peekable();
-        let mut previous_command = None;
-
-        while let Some(command) = commands.next()  {
-
-            let mut parts = command.trim().split_whitespace();
-            let command = parts.next().unwrap();
-            let args = parts;
-
-            match command {
-                "cd" => execute_cd(args.peekable()),
-                "exit" => return,
-                command => {
-                    let stdin = previous_command
-                        .map_or(
-                            Stdio::inherit(),
-                            |output: Child| Stdio::from(output.stdout.unwrap())
-                        );
-
-                    let stdout = if commands.peek().is_some() {
-                        // there is another command piped behind this one
-                        // prepare to send output to the next command
-                        Stdio::piped()
-                    } else {
-                        // there are no more built_ins piped behind this one
-                        // send output to shell stdout
-                        Stdio::inherit()
-                    };
-
-                    let output = Command::new(command)
-                        .args(args)
-                        .stdin(stdin)
-                        .stdout(stdout)
-                        .spawn();
-
-                    match output {
-                        Ok(output) => { previous_command = Some(output); },
-                        Err(e) => {
-                            previous_command = None;
-                            eprintln!("{}", e);
-                        },
-                    };
+            let readline = rl.readline(&*replace_placeholders(print_statement.as_str()));
+            match readline {
+                Ok(line) => {
+                    if line.trim() == "exit" {
+                        break;
+                    }
+                    rl.add_history_entry(line.as_str()).expect("TODO: Couldn't add to history");
+                    command_handler(line);
+                },
+                Err(ReadlineError::Interrupted) => {
+                    println!("CTRL-C");
+                    break
+                },
+                Err(ReadlineError::Eof) => {
+                    println!("CTRL-D");
+                    break
+                },
+                Err(err) => {
+                    println!("Error: {:?}", err);
+                    break
                 }
             }
         }
-
-        if let Some(mut final_command) = previous_command {
-            // block until the final command has finished
-            final_command.wait().expect("Fatal Error. Couldn't wait for command execution");
-        }
-
     }
+    rl.save_history(&dirs::home_dir().unwrap().join(".rsh_history")).expect("TODO: panic message");
+    Ok(())
 }
